+++ src/main/kotlin/com/example/dongguktoocean/cargo/repository/CargoRepository.kt
@@ -1,0 +1,6 @@
+package com.example.dongguktoocean.cargo.repository
+
+import com.example.dongguktoocean.cargo.domain.Cargo
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface CargoRepository : JpaRepository<Cargo, Long>
