+++ src/main/kotlin/com/example/dongguktoocean/document/repository/DocumentListRepository.kt
@@ -1,0 +1,6 @@
+package com.example.dongguktoocean.document.repository
+
+import com.example.dongguktoocean.document.domain.DocumentList
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface DocumentListRepository : JpaRepository<DocumentList, Long>
