+++ src/main/kotlin/com/example/dongguktoocean/document/repository/InvoiceDocRepository.kt
@@ -1,0 +1,6 @@
+package com.example.dongguktoocean.document.repository
+
+import com.example.dongguktoocean.document.domain.InvoiceDoc
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface InvoiceDocRepository : JpaRepository<InvoiceDoc, Long>
