--- src/main/kotlin/com/example/dongguktoocean/domain/Schedule.kt
@@ -1,30 +1,0 @@
-package com.example.dongguktoocean.domain
-
-import jakarta.persistence.*
-
-
-@Entity
-@Table(name = "Ship_Schedule")
-data class Schedule(
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    val id: Long? = null,
-
-    @Column(nullable = false)
-    val shipCode: String,
-
-    @Column(nullable = false)
-    val captain: String,
-
-    @Column(nullable = false)
-    val loadingPort: String,
-
-    @Column(nullable = false)
-    val destinationPort: String,
-
-    @Column(nullable = false)
-    val departureTime: String,
-
-    @Column(nullable = false)
-    val arrivalTime: String
-)
