--- src/main/kotlin/com/example/dongguktoocean/repository/CargoRepository.kt
@@ -1,6 +1,0 @@
-package com.example.dongguktoocean.repository
-
-import com.example.dongguktoocean.domain.Cargo
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface CargoRepository : JpaRepository<Cargo, Long>
