--- src/main/kotlin/com/example/dongguktoocean/repository/UserRepository.kt
@@ -1,7 +1,0 @@
-package com.example.dongguktoocean.repository
-
-
-import com.example.dongguktoocean.domain.User
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface UserRepository : JpaRepository<User, Long>
