+++ src/main/kotlin/com/example/dongguktoocean/user/repository/UserRepository.kt
@@ -1,0 +1,7 @@
+package com.example.dongguktoocean.user.repository
+
+
+import com.example.dongguktoocean.user.domain.User
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface UserRepository : JpaRepository<User, Long>
