+++ src/main/resources/mapper/ShipScheduleMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.example.dongguktoocean.repository.ShipScheduleRepository">
-    <select id="findSchedules" resultType="com.example.dongguktoocean.dto.ShipScheduleResponseDto">
+<mapper namespace="com.example.dongguktoocean.schedule.repository.ShipScheduleRepository">
+    <select id="findSchedules" resultType="com.example.dongguktoocean.schedule.dto.ShipScheduleResponseDto">
         SELECT
             ship_code AS shipCode,
             captain,
